Fix solvent module import hijacking
